+++ controllers/contacts-controller.js
@@ -1,4 +1,3 @@
-import Joi from "joi";
 import { HttpError } from "../helpers/index.js"
 import {
     listContacts,
@@ -10,43 +9,24 @@ import {
 import { ctrlWrapper } from "../decorators/index.js";
 
 
-const contactsSchema = Joi.object({
-    name: Joi.string().required().messages({
-        "any.required": `missing required name field`,
-    }),
-    email: Joi.string().required().messages({
-        "any.required": `missing required email field`,
-    }),
-    phone: Joi.string().required().messages({
-        "any.required": `missing required phone field`,
-    })
-}).required().messages({
-    "any.required": "missing fields",
-});
-
-const getAllContactsCtrl = async (req, res, next) => {
+const getAllContactsCtrl = async (_, res) => {
     const contacts = await listContacts();
     res.json(contacts);
 }
 
-const getOneContactCtrl = async (req, res, next) => {
+const getOneContactCtrl = async (req, res) => {
     const { contactId } = req.params;
     const result = await getContactById(contactId);
     if (!result) throw HttpError(404, `Movie with id=${id} not found`);
     res.json(result);
 }
 
-const addContactCtrl = async (req, res, next) => {
-
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, error.message);
-    }
+const addContactCtrl = async (req, res) => {
     const result = await addContact(req.body);
     res.status(201).json(result);
 }
 
-const delContactCtrl = async (req, res, next) => {
+const delContactCtrl = async (req, res) => {
     const { contactId } = req.params;
     const result = await removeContact(contactId);
     if (!result) {
@@ -59,12 +39,7 @@ const delContactCtrl = async (req, res, next) => {
     })
 }
 
-const putContactCtrl = async (req, res, next) => {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, error.message);
-    }
-
+const putContactCtrl = async (req, res) => {
     const { contactId } = req.params;
     const result = await updateContact(contactId, req.body);
