+++ routes/api/contacts.js
@@ -1,5 +1,8 @@
 import express from "express";
 import { contactsController } from "../../controllers/index.js";
+import validator from "../../decorators/validator.js";
+import contactsSchema from "../../schema/contacts-schema.js";
+
 
 const router = express.Router()
 
@@ -7,10 +10,10 @@ router.get('/', contactsController.getAllContactsCtrl)
 
 router.get('/:contactId', contactsController.getOneContactCtrl)
 
-router.post('/', contactsController.addContactCtrl)
-
 router.delete('/:contactId', contactsController.delContactCtrl)
 
-router.put('/:contactId', contactsController.putContactCtrl)
+router.post('/', validator(contactsSchema), contactsController.addContactCtrl)
+
+router.put('/:contactId', validator(contactsSchema), contactsController.putContactCtrl)
 
 export default router
