+++ schema/contacts-schema.js
@@ -1,0 +1,19 @@
+import Joi from "joi";
+
+const contactsSchema = Joi.object({
+    name: Joi.string().required().messages({
+        "any.required": `missing required name field`,
+    }),
+    email: Joi.string().required().messages({
+        "any.required": `missing required email field`,
+    }),
+    phone: Joi.string().required().messages({
+        "any.required": `missing required phone field`,
+    })
+}).required().messages({
+    "any.required": "missing fields",
+});
+
+export default {
+    contactsSchema,
+}
